requirements.txt
streamlit>=1.20
openai>=1.0.0
python-dotenv
PyPDF2
langchain>=0.0.278   # si sigues usando LangChain
scikit-learn
langgraph            # si sigues usando LangGraph
